Fix issue with non-available coins.
This was caused by the fact that any coin that had a hashdragon
op_return as output was being rejected;  in a wallet where all change
comes from hashdragon output, this is a problem.  So instead make sure
that there are vouts other than hashdragon owner vout.

Also added code for hibernating.  Code is just copy-paste, and needs
to refactored.